refactor: move process game into it's own helper file in /utils
