+++ app/picks/IPicks.ts
@@ -1,0 +1,8 @@
+import { Models } from 'appwrite/types/models';
+import { IGameWeek } from '@/api/IapiFunctions';
+
+export interface IWeeklyPicksProps {
+  NFLTeams: Models.Document[];
+  currentGameWeek: IGameWeek;
+  leagueId: string;
+}
